requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
brotli-python=1.0.9=py38h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cairo=1.16.0=hf32fb01_1
certifi=2023.11.17=py38h06a4308_0
cffi=1.16.0=py38h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cleverhans=4.0.0=pypi_0
cloudpickle=3.1.0=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py38h130f0dd_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
dm-tree=0.1.8=pypi_0
easydict=1.13=pypi_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.47.2=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.2=pypi_0
gast=0.6.0=pypi_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h5eee18b_3
glib=2.68.4=h9c3ff4c_1
glib-tools=2.68.4=h9c3ff4c_1
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
gpustat=1.1.1=pypi_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.18.5=h76c114f_0
harfbuzz=2.9.1=h83ec7ef_1
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.20.3=pypi_0
icu=68.1=h2531618_0
idna=3.4=py38h06a4308_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jasper=1.900.1=hd497a04_4
jinja2=3.1.3=py38h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h568e23c_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=5_h92ddd45_netlib
libclang=11.1.0=default_ha53f305_1
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.8.1.2=0
libcurand=10.3.4.107=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libevent=2.1.10=h9b69904_4
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.68.4=h174f98d_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=5_h92ddd45_netlib
liblapacke=3.9.0=5_h92ddd45_netlib
libllvm11=11.1.0=h9e868ea_6
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libopencv=4.4.0=py38_2
libpng=1.6.39=h5eee18b_0
libpq=12.15=h37d81fd_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.2.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.4=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mnist=0.2.2=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
mysql-common=8.0.25=ha770c72_2
mysql-libs=8.0.25=hfa10184_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
ninja=1.11.1.1=pypi_0
nose=1.3.7=pypi_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numpy=1.24.4=pypi_0
numpy-base=1.24.3=py38h060ed82_1
nvidia-ml-py=12.535.133=pypi_0
opencv=4.4.0=py38_2
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
pcre=8.45=h295c915_0
pillow=10.0.1=py38ha6cbd5a_0
pip=23.3.1=py38h06a4308_0
pixman=0.40.0=h7f8727e_1
psutil=5.9.8=pypi_0
py-opencv=4.4.0=py38h23f93f0_2
pycodestyle=2.12.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.18=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
qt=5.12.9=h9d6b050_2
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests>=2.31.0
safetensors=0.4.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorflow-probability=0.21.0=pypi_0
threadpoolctl=3.2.0=pypi_0
timm=0.4.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=pypi_0
torchaudio=2.0.0=py38_cu117
torchtriton=2.0.0=py38
torchvision=0.15.0=py38_cu117
tqdm=4.66.1=pypi_0
transformers=4.37.2=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.4=pypi_0
urllib3=1.26.18=py38h06a4308_0
wcwidth=0.2.13=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
